стили компонетнов вынесены в отдельные файлы
