+++ src/App.styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const App = styled.div`
+width: 960px;
+margin: auto;
+font-family: arial;
+`;
