+++ src/component/ImageList/ImageList.styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const ImageList = styled.div`
+display: flex;
+flex-direction: row;
+flex-wrap: wrap;
+`;
