+++ src/view/ImageInfo/ImageInfo.styles.ts
@@ -1,0 +1,23 @@
+import styled from 'styled-components'
+
+export const BackButton = styled.button`
+width: 100%;
+display: flex;
+flex-direction: row;
+justify-content: center;
+
+font-family: arial;
+font-size: 1.1em;
+
+padding: 16px 24px;
+
+border: none;
+outline: none;
+border-radius: 4px;
+color: white;
+background-color: #5C6BC0;
+
+margin-top: 16px;
+
+cursor: pointer;
+`;
