+++ src/view/ImageInfo/ImageInfo.tsx
@@ -5,35 +5,13 @@ import {
   useNavigate
 } from 'react-router-dom'
 
-import { ViewModel } from '../types/types';
+import { ViewModel } from '../../types/types';
 
-import NotFound from './NotFound';
+import NotFound from '../NotFound';
 
-import ImageInfoCard from '../component/ImageInfoCard'
+import ImageInfoCard from '../../component/ImageInfoCard/ImageInfoCard'
 
-import styled from 'styled-components'
-
-const BackButton = styled.button`
-width: 100%;
-display: flex;
-flex-direction: row;
-justify-content: center;
-
-font-family: arial;
-font-size: 1.1em;
-
-padding: 16px 24px;
-
-border: none;
-outline: none;
-border-radius: 4px;
-color: white;
-background-color: #5C6BC0;
-
-margin-top: 16px;
-
-cursor: pointer;
-`;
+import { BackButton } from './ImageInfo.styles'
 
 type ImageInfoProps = {
   viewModel: ViewModel
