+++ src/view/MainView/MainView.styles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const Main = styled.main`
+margin-top: 16px;
+`;
